PlayingWithNLP/requirements.txt
gensim == 4.1.2
google-cloud-language == 2.3.1